+++ app/src/main/java/org/sopt/official/domain/entity/notification/NotificationHistoryItem.kt
@@ -1,7 +1,7 @@
 package org.sopt.official.domain.entity.notification
 
 data class NotificationHistoryItem(
-    val notificationId: Int,
+    val notificationId: Long,
     val userId: Int,
     val title: String,
     val content: String?,
